Charts.tsx added. Clead up. Ability to draw diffrent charts.
